+++ SQLScripts/ClassSection_SP/delete_ClassSection SP.sql
@@ -4,8 +4,9 @@
 -- Description:	Procedure to Delete from the ClassSection Table.
 -- =============================================
 
-EXEC delete_ClassSection 3
+EXEC delete_ClassSection 7
 select * from ClassSection
+
 
 USE CalendarDB
 GO
@@ -16,21 +17,28 @@ IF EXISTS (SELECT * FROM sys.objects WHERE type = 'P' AND name = 'delete_ClassSe
 GO
 
 CREATE PROCEDURE delete_ClassSection(
-@ClassSectionID_1 int,
-@Msg nvarchar(MAX) = null OUTPUT
+@ClassSectionID_1 int
 )
 AS
-BEGIN TRY
+IF NOT EXISTS 
+(
+SELECT ClassSectionID
+FROM ClassSection
+WHERE ClassSectionID = @ClassSectionID_1)
+BEGIN  
+	-- Return 99 to the calling program to indicate failure.  
+	PRINT N'An error occurred deleting the class section information.';  
+	RETURN 99;  
+END
+ELSE 
 	DELETE FROM ClassSection
 	WHERE ClassSectionID = @ClassSectionID_1
-
-	SET @Msg = 'Class Section Deleted Successfully'
-
-END TRY
-BEGIN CATCH  
-    SELECT ERROR_MESSAGE() AS ErrorMessage;  
-END CATCH;  
-GO 
+    BEGIN  
+        -- Return 0 to the calling program to indicate success.  
+        PRINT N'The class section has been deleted.';  
+        RETURN 0;  
+    END;  
+GO
 
 -- Grant usage to the app user (needs to happen again when it gets deleted and re-created)
 GRANT EXECUTE ON [delete_ClassSection] TO appUserCalendarDB;
